requirements.txt
moviepy==1.0.3
numpy==2.1.3
openai==1.55.3
# numba_scipy==0.4.0
openai_whisper==20240930
Pillow==11.0.0 
llvmlite == 0.34.0
captacity==0.3.1
ffmpeg==1.4
pytube
pyttsx3
transformers